src/requirements.txt
google-genai
pyyaml
pydantic
python-dotenv
markdown
openai
anthropic
instructor
groq
python-frontmatter
cffi
cryptography
markitdown[all]